minilog: add linux specific build files for syslog helper
